Convert the character database from a list to a dictionary. The keys are traditional characters and the values are DbElem objects. DbElem objects can be scaled but for now there are three fields: traditional, simple, and index.
A traditional field must always be there. A simple field may be omitted.
An index must also be provided. This index will match the index of your
physical form to help you search up the characters quickly.

Updated the bare minimum to function with this new database. Searching
still works. And duplicates found on database init are still reported
correctly.

The next change will allow simplified and traditional characters to be
searched.