Update db.txt to 1361 characters. Made some corrections to the db as there were a few duplicates.
In terms of actual code changes, now you can search both traditional and
simplified character variants! Granted if there is a simplified version
in the db.txt, then the search should succeed.

Removed the dbElem class as it isn't needed right now. Modified the
initDb to initialize two different databases, one for traditional and
one for simplified. Searching is also modified. You can optionally
supply a traditional and/or simplified database. Obviously if you don't
supply any db the search will fail. Using a global variable to represent
the traditional and simplified databases. The databases map characters
to an index number.

The hope is in the future these characters can be referenced for more
detailed information. So these characters can be searched for their
index number, which in turn will be used to reference a specific
character's details.

Added some simplified variants to the db.txt. But still need to update a
lot more characters. Will slowly get to them over time. Be careful as
the characters are ASCII comma separated, not chinese input comma
separated!